Bot/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altgraph==0.17.4
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
packaging==23.2
pefile==2023.2.7
pyaes==1.6.1
pyasn1==0.6.0
pycparser==2.22
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PyJWT==2.8.0
pywin32-ctypes==0.2.2
requests==2.31.0
rsa==4.9
setuptools==69.0.3
Telethon==1.35.0
urllib3==2.2.1
yarl==1.9.4
